test: move fast-build wasm stub to fixture
